feat: moved settings drawer into video view
