--- libs/settings-drawer/.eslintrc.json
@@ -1,18 +1,0 @@
-{
-  "extends": ["../../.eslintrc.json"],
-  "ignorePatterns": ["!**/*"],
-  "overrides": [
-    {
-      "files": ["*.ts", "*.js"],
-      "rules": {}
-    },
-    {
-      "files": ["*.ts", "*.js"],
-      "rules": {}
-    },
-    {
-      "files": ["*.ts", "*.js"],
-      "rules": {}
-    }
-  ]
-}
