--- libs/settings-drawer/jest.config.ts
@@ -1,15 +1,0 @@
-/* eslint-disable */
-export default {
-  displayName: 'settings-drawer',
-  preset: '../../jest.preset.js',
-  globals: {
-    'ts-jest': {
-      tsconfig: '<rootDir>/tsconfig.spec.json',
-    },
-  },
-  transform: {
-    '^.+\\.[tj]sx?$': 'ts-jest',
-  },
-  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
-  coverageDirectory: '../../coverage/libs/settings-drawer',
-};
