--- libs/settings-drawer/project.json
@@ -1,24 +1,0 @@
-{
-  "name": "settings-drawer",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/settings-drawer/src",
-  "projectType": "library",
-  "tags": [],
-  "targets": {
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": ["libs/settings-drawer/**/*.{ts,js}"]
-      }
-    },
-    "test": {
-      "executor": "@nrwl/jest:jest",
-      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
-      "options": {
-        "jestConfig": "libs/settings-drawer/jest.config.ts",
-        "passWithNoTests": true
-      }
-    }
-  }
-}
